+++ articles/urls.py
@@ -15,5 +15,10 @@ urlpatterns = [
         autocomplete.PageAutocomplete.as_view(),
         name='page-autocomplete',
     ),
+    url(
+        r'^search/$',
+        views.ArticleSearchResultView.as_view(),
+        name='search',
+    ),
     url(r'^(?P<slug>[^/]+)/$', views.ArticleDetailView.as_view(), name='detail'),
 ]